Duplicator added. Documentation of static methods in SandLab.java improved

+++ sandbox/SandDisplay.java
@@ -68,7 +68,7 @@ public class SandDisplay extends JComponent implements MouseListener,
     slider.setPaintTicks(true);
     Hashtable<Integer, JLabel> labelTable = new Hashtable<Integer, JLabel>();
     labelTable.put(new Integer(0), new JLabel("Slow"));
-    labelTable.put(new Integer(100), new JLabel("Fast"));
+    labelTable.put(new Integer(80), new JLabel("Fast"));
     slider.setLabelTable(labelTable);
     slider.setPaintLabels(true);
